src/deployment_scripts/remove_services_from_stack/requirements.txt
pyyaml
environs
